+++ src/OpenTelemetry/CHANGELOG.md
@@ -17,6 +17,10 @@
   ([#5004](https://github.com/open-telemetry/opentelemetry-dotnet/pull/5004))
   ([#5016](https://github.com/open-telemetry/opentelemetry-dotnet/pull/5016))
 
+* Update `AggregatorStore` to provide known connection metrics with larger
+  histogram buckets.
+  ([#5008](https://github.com/open-telemetry/opentelemetry-dotnet/pull/5008))
+
 ## 1.7.0-alpha.1
 
 Released 2023-Oct-16